Update character count to work with wrapped inputs
If the character count component is used on a text input, that contains a prefix or suffix the textareaDescription is incorrectly placed.

As it gets inserted immediately after the $textarea element (input) this means it is within the input wrapper flex container.

Check for the presence of a `.govuk-input__wrapper` element, and if it exists, insert the description after the wrapper element.